projects/backend-GPTvision-2/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
requests>=2.31.0
pillow>=10.3.0
openai>=1.0.0
python-dotenv>=1.0.0 